# swagger.yaml
openapi: 3.0.0
info:
  title: Distributed Parcel Delivery & Tracking System
  version: 1.0.1
  description: |
    OpenAPI 3.0 spec updated to reflect implemented backend endpoints.
servers:
  - url: http://localhost:3001
    description: user-service (auth, users)
  - url: http://localhost:3002
    description: order-service
  - url: http://localhost:3003
    description: payment-service
...
paths:
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user (customer or courier)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '400': { description: Bad request }
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login and receive token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401': { description: Invalid credentials }

  # User endpoints (mounted under /api/users)
...
  /api/orders:
    post:
      tags: [Orders]
      summary: Create order (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      tags: [Orders]
      summary: Get orders (filters available)
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string }
        - name: customerId
          in: query
          schema: { type: string }
        - name: courierId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Orders list with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
                  pagination:
                    type: object
...
components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name, phone_number]
      properties:
        email: { type: string }
        password: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        phone_number: { type: string }
        user_role:
          type: string
          enum: [CUSTOMER, COURIER]
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    User:
      type: object
      properties:
        user_id: { type: string }
        email: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        phone_number: { type: string }
        user_role: { type: string }
...
    Address:
      type: object
      properties:
        address_id: { type: string }
        address_type: { type: string }
        street_address: { type: string }
        subcity: { type: string }
        kebele: { type: string }
        house_number: { type: string }
        landmark: { type: string }
        instructions: { type: string }
    AddressCreate:
      type: object
      required: [address_type, street_address]
      properties:
        address_type: { type: string }
        street_address: { type: string }
        contact_name: { type: string }
        contact_phone: { type: string }
...
    OrderCreateRequest:
      type: object
      required: [addresses, parcels]
      properties:
        priority: { type: string }
        serviceType: { type: string }
        scheduledPickupTime: { type: string, format: date-time }
        scheduledDeliveryTime: { type: string, format: date-time }
        notes: { type: string }
        addresses:
          type: array
          items: { $ref: '#/components/schemas/AddressCreate' }
        parcels:
          type: array
          items:
            type: object
            properties:
              description: { type: string }
              weightKg: { type: number }
              lengthCm: { type: number }
              widthCm: { type: number }
              heightCm: { type: number }
              declaredValue: { type: number }
              category: { type: string }
              isFragile: { type: boolean }
              isPerishable: { type: boolean }
              requiresSignature: { type: boolean }
    Order:
      type: object
      properties:
        orderId: { type: string }
        orderNumber: { type: string }
        customerId: { type: string }
        status: { type: string }
        priority: { type: string }
        serviceType: { type: string }
        price: { type: number }
        addresses:
          type: array
          items: { $ref: '#/components/schemas/Address' }
        parcels:
          type: array
          items:
            type: object
            properties:
              parcelNumber: { type: string }
              description: { type: string }
              weightKg: { type: number }